# aulas/05-foundry-agents/Lab2/agente/openapi-agente.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Serviço de Continuidade do Deva",
    "description": "Memória e fila do agente Deva. O agente LÊ a memória, LÊ a fila, AVANÇA documentos e PROPÕE regras novas. Aprovar propostas e liberar exceções são operações de auditor humano e não fazem parte desta especificação de propósito.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://SUA-URL"
    }
  ],
  "paths": {
...
    "/memoria/proposta": {
      "post": {
        "tags": [
          "memória"
        ],
        "summary": "Propor uma regra nova",
        "description": "O agente chama isto quando o auditor o corrige. A regra NÃO passa a valer: entra numa fila de revisão.",
        "operationId": "propor_memoria_proposta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaDeProposta"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropostaDeMemoria"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntradaDeProposta": {
        "properties": {
          "secao": {
            "$ref": "#/components/schemas/SecaoDaMemoria"
          },
          "texto": {
            "type": "string",
            "maxLength": 400,
            "minLength": 15,
            "title": "Texto"
          },
          "evidencia": {
            "type": "string",
            "maxLength": 400,
            "minLength": 10,
            "title": "Evidencia"
          }
        },
        "type": "object",
        "required": [
          "secao",
          "texto",
          "evidencia"
        ],
        "title": "EntradaDeProposta",
        "description": "O que o agente envia. Repare no que NÃO existe aqui: situação, aprovador, data\nde decisão. O agente não tem como se auto-aprovar porque o contrato não permite."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropostaDeMemoria": {
        "properties": {
          "identificador": {
            "type": "string",
            "title": "Identificador"
          },
          "secao": {
            "$ref": "#/components/schemas/SecaoDaMemoria"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "evidencia": {
            "type": "string",
            "title": "Evidencia",
            "description": "Qual documento e qual correção do auditor motivaram esta proposta."
          },
          "proposta_por": {
            "type": "string",
            "title": "Proposta Por",
            "default": "deva"
          },
          "proposta_em": {
            "type": "string",
            "format": "date-time",
            "title": "Proposta Em"
          },
          "situacao": {
            "$ref": "#/components/schemas/SituacaoDaProposta",
            "default": "pendente"
          },
          "decidida_por": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decidida Por"
          },
          "decidida_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decidida Em"
          },
          "motivo_do_descarte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Do Descarte"
          }
        },
        "type": "object",
        "required": [
          "identificador",
          "secao",
          "texto",
          "evidencia",
          "proposta_em"
        ],
        "title": "PropostaDeMemoria",
        "description": "Uma regra que o agente QUER aprender. Ainda não vale nada.\n\nO campo `evidencia` é o que separa aprendizado de alucinação: se o agente não\nconsegue apontar o documento e a conversa que motivaram a regra, a proposta não\ndeveria existir."
      },
...
      "SecaoDaMemoria": {
        "type": "string",
        "enum": [
          "classificacao",
          "interpretacao_de_politica",
          "fornecedores",
          "operacao"
        ],
        "title": "SecaoDaMemoria",
        "description": "As seções do MEMORY.md. Lista fechada de propósito.\n\nMemória sem seção vira lista infinita, e lista infinita ninguém revisa."
      },
      "SituacaoDaProposta": {
        "type": "string",
        "enum": [
          "pendente",
          "aprovada",
          "descartada"
        ],
        "title": "SituacaoDaProposta"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}